feat: add smooth order verification
