+++ src/loaders/index.ts
@@ -1,6 +1,7 @@
 export { loader as appProductsLoader } from "./app-products-loader";
 export { HydrateFallback } from "./hydrate-fallback";
 export { loader as instructionLoader } from "./instruction-loader";
+export { loader as orderStatusLoader } from "./order-status-loader";
 export { loader as productKitLoader } from "./product-kit-loader";
 export { loader as selectedKitLoader } from "./selected-kit-loader";
 export { loader as successfulScreenLoader } from "./successful-screen-loader";
